updated query id and qualifiers params in /users in removal of sensitive info(phone,email,tokens,chaincode) on /users route
