+++ controllers/users.js
@@ -8,10 +8,9 @@ const { logType } = require("../constants/logs");
 const dataAccess = require("../services/dataAccessLayer");
 const logger = require("../utils/logger");
 const obfuscate = require("../utils/obfuscate");
-const { getQualifiers } = require("../utils/helper");
-const { getUsernamesFromPRs } = require("../utils/users");
+// const { getUsernamesFromPRs } = require("../utils/users");
 const { SOMETHING_WENT_WRONG, INTERNAL_SERVER_ERROR } = require("../constants/errorMessages");
-const { getFilteredPRsOrIssues } = require("../utils/pullRequests");
+// const { getFilteredPRsOrIssues } = require("../utils/pullRequests");
 const { setInDiscordFalseScript } = require("../services/discordService");
 const { generateDiscordProfileImageUrl } = require("../utils/discord-actions");
 const { addRoleToUser, getDiscordMembers } = require("../services/discordService");
@@ -73,59 +72,7 @@ const getUserById = async (req, res) => {
  * @param res {Object} - Express response object
  */
 
-const removePersonalDetails = (user) => {
-  const { phone, email, ...safeUser } = user;
-  return safeUser;
-};
-
-/**
- * Fetches the data about our users
- *
- * @param req {Object} - Express request object
- * @param res {Object} - Express response object
- */
-
 const getUsers = async (req, res) => {
-  // getting user details by id if present.
-  const query = req.query?.query ?? "";
-  const qualifiers = getQualifiers(query);
-
-  // getting user details by id if present.
-  if (req.query.id) {
-    const id = req.query.id;
-    let result;
-    try {
-      result = await userQuery.fetchUser({ userId: id });
-    } catch (error) {
-      logger.error(`Error while fetching user: ${error}`);
-      return res.boom.serverUnavailable(SOMETHING_WENT_WRONG);
-    }
-
-    if (!result.userExists) {
-      return res.boom.notFound("User doesn't exist");
-    }
-
-    const User = { ...result.user };
-    const user = removePersonalDetails(User);
-
-    return res.json({
-      message: "User returned successfully!",
-      user,
-    });
-  }
-
-  if (qualifiers?.filterBy) {
-    const allPRs = await getFilteredPRsOrIssues(qualifiers);
-
-    const usernames = getUsernamesFromPRs(allPRs);
-
-    const { users } = await userQuery.fetchUsers(usernames);
-
-    return res.json({
-      message: "Users returned successfully!",
-      users,
-    });
-  }
   return dataAccess.retrieveUsers(req, res);
 };
